converting to css modules from App.css
- Moving css codes into component modules
- Responsive adjustment for Home, NavBar, Signup and Login
