+++ src/components/Home/Home.jsx
@@ -1,11 +1,11 @@
-import background from '../images/bg3.jpeg'
+import background from '../../images/bg3.jpeg'
+import styles from './Home.module.css'
 
 const Home = () => {
 	return (
 		<div className="home-page">
-			{/* <h1 >Cocktails Tell Stories</h1>  */}
 			<img
-				id="background-img"
+				className={styles.img}
 				src={background}
 				alt="background"
 				loading="lazy"
