+++ src/components/Home/Home.module.css
@@ -1,0 +1,10 @@
+.img {
+    background-size: cover;
+        background-repeat: no-repeat;
+        position: fixed;
+        top: 0;
+        left: 0px;
+        z-index: -10;
+        height: 100vh;
+        width: 100vw;
+}
