+++ src/features/auth/Login/Login.module.css
@@ -1,0 +1,126 @@
+.container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width: 100vw;
+}
+
+.form {
+    display: flex;
+    flex-direction: column;
+    gap: 1vh;
+    align-items: center;
+    opacity: 85%;
+    background: white;
+    padding: 0 80px 12px;
+    border-radius: 20px;
+    box-shadow: 10px 10px 10px 1px rgba(0, 0, 0, 0.2);
+}
+
+.form h1 {
+    width: 100%;
+    border-bottom: 2px solid black;
+    line-height: 40px;
+    width: 100%;
+    font-weight: bolder;
+    text-align: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 5vh;
+}
+
+.form input {
+    border: 1px solid lightgrey;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 20px;
+    padding: 4px 8px;
+    letter-spacing: 1px;
+    font-weight: 600;
+    margin: 0px 4px 0px;
+    font-size: 11pt;
+    background: transparent;
+    transition: 1s;
+    outline: none;
+}
+
+.form p {
+    color: #f20a0a;
+    font-size: 14px;
+    margin: 4px 0 0;
+    padding: 0;
+}
+
+.button {
+    width: 75%;
+    justify-content: center;
+    align-items: center;
+    background-color: #706897;
+    color: white;
+    margin-top: 5px;
+    border-radius: 20px;
+    border: none;
+    font-size: 15pt;
+    font-weight: 700;
+    transition: 1s;
+    outline: none;
+    cursor: pointer;
+    margin-bottom: 20px;
+}
+
+.button:hover {
+    transform: scale(1.03);
+    letter-spacing: 4px;
+    background-color: #7FFFD4;
+    color: #1E90FF;
+}
+
+@media (max-width: 390px) {
+    .form {
+        margin-top: 0;
+        padding: 0 12vw;
+    }
+    .form h1 {
+        font-size: 20px;
+    }
+    .button {
+        font-size: 16px;
+        width: 50%;
+        padding: 4px 0;
+    }
+}
+
+@media (min-width: 400px) and (max-width: 500px) {
+    .form {
+        margin-top: 0;
+        padding: 0 15vw;
+    }
+    .form h1 {
+        line-height: 40px;
+        font-size: 24px
+    }
+    .button {
+        font-size: 16px;
+        width: 60%;
+        padding: 4px 0;
+    }
+}
+
+@media (orientation: landscape) and (max-width: 800px) {
+    .form {
+        margin-top: 24px;
+        padding: 0 60px;
+    }
+    .form h1 {
+        line-height: 40px;
+        font-size: 24px
+    }
+    .button {
+        font-size: 16px;
+        width: 60%;
+        padding: 4px 0;
+    }
+}
