+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
 import { BrowserRouter as Router } from 'react-router-dom'
 import store, { persistor } from './app/store'
-import App from './components/App'
+import App from './components/App/App'
 import { categoriesApi } from './features/categories/categoriesSlice'
 import { usersApi } from './features/users/usersSlice'
 import './index.css'
